requirements.txt
# Core ML — torch already installed system-wide, verify: python -c "import torch; print(torch.__version__)"
torch>=2.2.0
torchvision>=0.17.0
transformers>=4.45.0
accelerate>=0.30.0
qwen-vl-utils>=0.0.8

# Optional: 4-bit quantization for Qwen on smaller GPUs
# bitsandbytes>=0.43.0

# SAM2 — install separately:
#   pip install git+https://github.com/facebookresearch/sam2.git
# SAM1 fallback:
#   pip install segment-anything

# Image processing
Pillow>=10.0.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
scipy>=1.11.0

# COCO utilities
pycocotools>=2.0.7

# API
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9

# GUI
gradio>=4.36.0

# Config & utilities
pyyaml>=6.0.1
tqdm>=4.66.0
aiofiles>=23.2.0
